# web/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tamandua",
    "version": "0.9"
  },
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_dashboard_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "tz",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tz"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/audit": {
      "post": {
        "tags": [
          "dashboard"
        ],
        "summary": "Audit Report",
        "description": "Audit evidence of a run or an asset's state, with the chosen findings (`fingerprints`; all of the scope without it).\n\nSame access as seeing that run or asset.",
        "operationId": "audit_report_api_reports_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "run_id": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^[0-9a-f]{32}$"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Run Id"
                  },
                  "asset": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 200
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Asset"
                  },
                  "account": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 100
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Account"
                  },
                  "assets": {
                    "anyOf": [
                      {
                        "items": {
                          "type": "string"
                        },
                        "type": "array",
                        "maxItems": 500,
                        "minItems": 1
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Assets"
                  },
                  "status": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "open",
                          "fixed",
                          "all"
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Status"
                  },
                  "fingerprints": {
                    "anyOf": [
                      {
                        "items": {
                          "type": "string"
                        },
                        "type": "array",
                        "maxItems": 5000
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Fingerprints"
                  },
                  "options": {
                    "anyOf": [
                      {
                        "additionalProperties": true,
                        "type": "object"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Options"
                  }
                },
                "additionalProperties": false,
                "type": "object",
                "title": "AuditReportIn",
                "description": "Exactly one scope: a run, an asset's state, an organization (`account`) or a selection of assets."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Audit evidence (PDF)",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "maxLength": 50000000,
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Dashboard": {
        "properties": {
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "kpis": {
            "additionalProperties": true,
            "type": "object",
            "title": "Kpis"
          },
          "issues_over_time": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 366,
            "title": "Issues Over Time"
          },
          "open_vs_fixed": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 366,
            "title": "Open Vs Fixed"
          },
          "top_assets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Top Assets"
          },
          "by_cwe": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 8,
            "title": "By Cwe"
          },
          "exploitability": {
            "additionalProperties": true,
            "type": "object",
            "title": "Exploitability"
          },
          "activity": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 365,
            "title": "Activity"
          },
          "recent_runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Recent Runs"
          },
          "top_issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Top Issues"
          },
          "kev_news": {
            "additionalProperties": true,
            "type": "object",
            "title": "Kev News"
          },
          "cve_news": {
            "additionalProperties": true,
            "type": "object",
            "title": "Cve News"
          },
          "tools": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "window_days",
          "generated_at",
          "kpis",
          "issues_over_time",
          "open_vs_fixed",
          "top_assets",
          "by_cwe",
          "exploitability",
          "activity",
          "recent_runs",
          "top_issues",
          "kev_news",
          "cve_news",
          "tools"
        ],
        "title": "Dashboard"
      },
...
      "Error": {
        "title": "Error",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "In the reader's language."
          }
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tamandua_session",
        "description": "Session signed in to the panel."
      },
      "metrics": {
        "type": "http",
        "scheme": "bearer",
        "description": "TAMANDUA_METRICS_TOKEN (Prometheus scraper)."
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}